Fix test that started failing after SDK upgrade
This was working before only due to an unintended interaction between HCL
and HIL, where HCL didn't understand that $${ is an _escaped_
interpolation sequence and so activated its brace-counting closing quote
exception even though the quotes were not actually inside an interpolation
sequence.

With HCL and HIL now merged together in HCL2, the scanner knows that these
nested quotes are _not_ inside an interpolation sequence and so they need
to be escaped to avoid terminating the string literal early.

Terraform 0.12's configuration upgrade tool will fix this up for normal
user configurations because it processes the input using HCL/HIL and then
re-applies canonical escaping to the resulting string literals, which
therefore causes these quotes to get escaped as a side-effect.

This also adjusts the test implementation slightly to use t.Run, which
means we'll complete testing all of the cases even if one of them fails.
The test logic is otherwise unchanged.