Use HCL2 templates instead of HIL templates
This is broadly compatible but introduces support for loops and
conditional blocks along with the existing support for simple
interpolations.

With that comes some small incompatibilities, though. Most notably, any
template containing the control sequence introducer %{ will need to be
rewritten to escape it as %%{ .

This also includes a heavy revision of the docs to essentially dissade use
of template_file for those using Terraform 0.12 or later, because it now
has template file rendering available as a built-in function.