template_cloudinit_config: Require base64 when gzip is enabled
In Terraform 0.11 we would sometimes silently corrupt binary data in
strings since data traverses through representations that only support
unicode strings, like JSON files.

In Terraform 0.12 we're checking this more strongly to avoid that sort of
silent corruption, and so unencoded gzip will get trapped explicitly, but
adding this explicit check in the provider allows us to report the error
in a more contextual way rather than flagging it as a provider bug.