Ignore missing paths during Read phase
When the source directory can't be found during a Read, it isn't
meaningful to try to hash it and compare the hashes. Instead we should
mark it dirty and continue to the Create step, where we will load the
source templates based on the current module location instead of the
plan contents.

Fixes terraform-providers/terraform-provider-template#19